requirements.txt
streamlit
pandas
numpy
scikit-learn 
pdf2image 
google-generativeai 
python-dotenv